% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdaycard.R
\name{bday_color}
\alias{bday_color}
\title{Birthday Card Colors}
\usage{
bday_color(
  cakebottom,
  caketop,
  candle1,
  candle2,
  candle3,
  candle4,
  candle5,
  candle6,
  line1,
  line2,
  line3,
  star1,
  star2
)
}
\arguments{
\item{cakebottom}{cake bottom color}

\item{caketop}{cake top color}

\item{candle1}{candle1 color}

\item{candle2}{candle2 color}

\item{candle3}{candle3 color}

\item{candle4}{candle4 color}

\item{candle5}{candle5 color}

\item{candle6}{candle6 color}

\item{line1}{line1 color}

\item{line2}{line2 color}

\item{line3}{line3 color}

\item{star1}{star1 color}

\item{star2}{star2 color}
}
\value{
the colors for the birthday card
}
\description{
Birthday Card Colors
}
