% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday_card.R
\docType{data}
\name{birthday_data}
\alias{birthday_data}
\title{Birthday_card Data}
\format{
A table with 6 columns, which are:
\describe{
\item{x}{the position of the points on there x-axis}
\item{y}{the position of the points on there y-axis}
\item{group}{the group of the point}
\item{size}{the size of the point}
\item{t}{the thansition step of the point}
\item{shape}{the shape of the point}
}
}
\usage{
birthday_data
}
\description{
The data for plotting the card
}
\keyword{datasets}
