% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdaycard.R
\name{birthday_video}
\alias{birthday_video}
\title{Birthday Card Video}
\usage{
birthday_video(
  birthdaycard,
  audio = "b_day_music.mp3",
  duration = 12,
  end_pause = 2
)
}
\arguments{
\item{birthdaycard}{the GIF birthday card witch created by \code{birthdaycard}}

\item{audio}{the audio play with the video}

\item{duration}{the duration of the video}

\item{end_pause}{the pause time at the end}
}
\value{
A video format of birthday card
}
\description{
Birthday Card Video
}
