% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_card.R
\docType{data}
\name{christmas_data}
\alias{christmas_data}
\title{Christmas Card Data}
\format{
A table with 5 columns, which are:
\describe{
\item{x}{the position of the points on there x-axis}
\item{y}{the position of the points on there y-axis}
\item{group}{the group of the point}
\item{size}{the size of the point}
\item{t}{the thansition step of the point}
}
}
\usage{
christmas_data
}
\description{
The data for plotting the Christmas card
}
\keyword{datasets}
