% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read.R
\name{data_read}
\alias{data_read}
\title{Read CSV duplicate}
\usage{
data_read(file)
}
\arguments{
\item{file}{path to file name}
}
\value{
a table
}
\description{
Read CSV duplicate
}
\examples{
birthday_data = system.file("extdata", "birthday_data.csv", package="grcard")
data_read(birthday_data)
}
