% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_card.R
\name{xmas_color}
\alias{xmas_color}
\title{Christmas Card Colors}
\usage{
xmas_color(
  ball1,
  ball2,
  ball3,
  ball4,
  ball5,
  ball6,
  ball7,
  star,
  treebottom,
  treelower,
  treeupper,
  treetop,
  trunk
)
}
\arguments{
\item{ball1}{ball color}

\item{ball2}{ball color}

\item{ball3}{ball color}

\item{ball4}{ball color}

\item{ball5}{ball color}

\item{ball6}{ball color}

\item{ball7}{ball color}

\item{star}{star color}

\item{treebottom}{treebottom color}

\item{treelower}{treelower color}

\item{treeupper}{treeupper color}

\item{treetop}{treetop color}

\item{trunk}{trunk color}
}
\value{
a list of colors for the Christmas card
}
\description{
Christmas Card Colors
}
