% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_card.R
\name{xmas_video}
\alias{xmas_video}
\title{Christmas Video}
\usage{
xmas_video(
  christmascard,
  audio = "xmas music.mp3",
  duration = 13,
  end_pause = 2
)
}
\arguments{
\item{christmascard}{the GIF Christmas card}

\item{audio}{the audio play with the video}

\item{duration}{the duration of the video}

\item{end_pause}{the pause time at the end}
}
\value{
A video format of birthday card
}
\description{
Christmas Video
}
