% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday_card.R
\name{bday_card}
\alias{bday_card}
\title{Birthday Card GIF}
\usage{
bday_card(greeting = "Happy Birthday", color = "cream")
}
\arguments{
\item{greeting}{the greeting words to show on the card, default is "Happy Birthday".}

\item{color}{the color of the birthday card, could be choose from "cream",
"chocolate" and "strawberry",default is "cream"}
}
\value{
A GIF format birthday card
}
\description{
Create a birthday card in GIF format
}
\examples{
bday_card()
}
