% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthdaycard.R
\name{bday_video}
\alias{bday_video}
\title{Birthday Card Video}
\usage{
bday_video(birthdaycard, audio = bday_music_mp3, duration = 12, end_pause = 2)
}
\arguments{
\item{birthdaycard}{the GIF birthday card witch created by \code{birthdaycard}}

\item{audio}{the audio play with the video}

\item{duration}{the duration of the video}

\item{end_pause}{the pause time at the end}
}
\value{
A video format of birthday card
}
\description{
Birthday Card Video
}
\examples{
birthdaycard<-bday_card()
bday_video(birthdaycard)
}
