% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_card.R
\name{xmas_card}
\alias{xmas_card}
\title{Christmas Card GIF}
\usage{
xmas_card(greeting = "Merry Christmas", color = "candy")
}
\arguments{
\item{greeting}{the greeting words to show on the card, defult is "Merry Christmas"}

\item{color}{the color of the Christmas card, could be choose from "forest",
"candy" and "dream", or create your own color combination by using. Defult is "candy"
the \code{xmas_color}}
}
\value{
A GIF format Christmas card
}
\description{
Create a Christmas card in GIF format
}
\examples{
xmas_card()
}
