% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christmas_card.R
\name{xmas_video}
\alias{xmas_video}
\title{Christmas Card Video}
\usage{
xmas_video(christmascard, audio = xmas_music_mp3, duration = 12, end_pause = 2)
}
\arguments{
\item{christmascard}{the GIF Christmas card witch created by \code{xmas_card}}

\item{audio}{the audio play with the video}

\item{duration}{the duration of the video}

\item{end_pause}{the pause time at the end}
}
\value{
A video format of Christmas card
}
\description{
Christmas Card Video
}
\examples{
christmascard<-xmas_card(color="candy")
xmas_video(christmascard)
}
